+++ room-service/src/main/java/soa/roomservice/dto/RoomResponse.java
@@ -16,6 +16,6 @@ public class RoomResponse {
     private String roomNumber;
     private String description;
     private Integer capacity;
-    private int cost;
+    private Double cost;
     private RoomType type;
 }
